Added Year To Date to SubmissionSummary
