+++ src/SFA.DAS.Payments.Monitoring.Metrics.Model/Submission/SubmissionSummaryModel.cs
@@ -10,6 +10,7 @@ namespace SFA.DAS.Payments.Monitoring.Metrics.Model.Submission
         public short AcademicYear { get; set; }
         public byte CollectionPeriod { get; set; }
         public decimal Percentage { get; set; }
+        public ContractTypeAmounts YearToDatePayments { get; set; }
         public ContractTypeAmountsVerbose SubmissionMetrics { get; set; }
         public virtual List<EarningsModel> EarningsMetrics { get; set; }
         public virtual List<DataLockedEarningsModel> DataLockedPaymentsMetrics { get; set; }
